secondo aggiornamento

+++ src/main/java/com/example/demo/service/DocenteService.java
@@ -1,6 +1,9 @@
 package com.example.demo.service;
 import com.example.demo.DTO.DocenteDTO;
+import com.example.demo.DTO.DocenteDTOFormat;
+import com.example.demo.entity.Corso;
 import com.example.demo.entity.Docente;
+import com.example.demo.repository.CorsoRepository;
 import com.example.demo.repository.DocenteRepository;
 import com.example.demo.utils.DocenteConverter;
 import jakarta.persistence.EntityNotFoundException;
@@ -15,19 +18,22 @@ import java.util.Optional;
 public class DocenteService {
 
     private final DocenteRepository docenteRepository;
-    public DocenteService(DocenteRepository docenteRepository)
+    private final CorsoRepository corsoRepository;
+
+    public DocenteService(DocenteRepository docenteRepository, CorsoRepository corsoRepository)
     {
         this.docenteRepository = docenteRepository;
+        this.corsoRepository = corsoRepository;
     }
 
     //READ
-    public DocenteDTO getDocenteById(Integer id)
+    public DocenteDTOFormat getDocente(Integer id)
     {
         Optional <Docente> docente= docenteRepository.findById(id);
         if(docente.isPresent())
         {
-            DocenteDTO docenteDTO = DocenteConverter.entityToDTO(docente.get());
-            return docenteDTO;
+            DocenteDTOFormat docenteDTOFormat = DocenteConverter.entityToDTOFormat(docente.get());
+            return docenteDTOFormat;
         }
         else
         {
@@ -36,35 +42,37 @@ public class DocenteService {
     }
 
     //READ ALL
-    public List<DocenteDTO>getAllDocenti()
+    public List<DocenteDTOFormat>getAllDocenti()
     {
         List<Docente> listaDocenti = docenteRepository.findAll();
-        List<DocenteDTO> listaDocentiDTO=new ArrayList<DocenteDTO>();
+        List<DocenteDTOFormat> listaDocentiDTOFormat=new ArrayList<DocenteDTOFormat>();
 
         for (Docente docente : listaDocenti)
         {
-            DocenteDTO docenteDTO = DocenteConverter.entityToDTO(docente);
-            listaDocentiDTO.add(docenteDTO);
+            DocenteDTOFormat docenteDTOFormat = DocenteConverter.entityToDTOFormat(docente);
+            listaDocentiDTOFormat.add(docenteDTOFormat);
         }
-        return listaDocentiDTO;
+        return listaDocentiDTOFormat;
     }
 
     //CREATE
-    public DocenteDTO insertDocente(DocenteDTO docenteDTO)
+    public DocenteDTO createDocente(DocenteDTO docenteDTO)
     {
         Docente docente = DocenteConverter.DTOToEntity(docenteDTO);
         docenteRepository.save(docente);
         return DocenteConverter.entityToDTO(docente);
     }
+
     //UPDATE
-    public DocenteDTO updateDocenteById(Integer id, DocenteDTO docenteDTO)
+    public DocenteDTOFormat updateDocente(Integer id, DocenteDTO docenteDTO)
     {
         Optional <Docente> docente = docenteRepository.findById(id);
         if(docente.isPresent()) {
             docenteDTO.setId(id);
             Docente docenteModificato = DocenteConverter.DTOToEntity(docenteDTO);
+            docenteModificato.setListaCorsi(docente.get().getListaCorsi());
             docenteRepository.save(docenteModificato);
-            return DocenteConverter.entityToDTO(docenteModificato);
+            return DocenteConverter.entityToDTOFormat(docenteModificato);
         }
         else
         {
@@ -72,11 +80,24 @@ public class DocenteService {
         }
     }
 
-    public void deleteDocenteById(Integer id)
+    //DELETE
+    public void deleteDocente(Integer id)
     {
         if (docenteRepository.existsById(id))
         {
-            docenteRepository.deleteById(id);
+            Docente docente = docenteRepository.findById(id).get();
+            List<Corso> corsi = corsoRepository.findAll();
+            for(Corso corso : corsi)
+            {
+                if (corso.getDocente().equals(docente))
+                {
+                    throw new EntityNotFoundException("Il docente è associato al corso id:"+ corso.getId() + " " + corso.getNomeCorso());
+                }
+                else
+                {
+                    docenteRepository.deleteById(id);
+                }
+            }
         }
         else
         {
